refactor(logic): split intervention_resolver into parts (Refs #2288)
Extract call-to-arms flow and apply-intervention helpers into library parts
so the primary file stays focused and under the 700-line threshold now that
intervention_resolver.dart grew past it. No behavior change; same public API.
